Move http tests into rest suite
Also requires changes to Http.java itself to allow its tests to be in a
different package.
